Feature: Meta Regressor (#6)
* Bugfix (preprocessing): Exclude boolean columns from imputation

As the boolean columns were not considered as categorical columns, the
value imputator treated them as numerical columns, as everything that was not
categorical was implicitly numerical. Now, the imputer recognized the third state
of a column being boolean.

* Feature (regression): Add meta regressor

The meta regressor fits a number of different linear and non-linear regressors
to the data. The resulting predictions of these regressors are then used as additional input
for a Lasso meta model.

* Feature (regression): Add stacked meta regression model

* Refactor (scripts): Remove prototype training script